category controller

+++ backend/service/CategoryService.go
@@ -8,11 +8,11 @@ import (
 
 type CategoryService interface {
 	NewCategory(category *model.Category) error
-	FindById(id int) (*model.Category, error)
+	FindById(id string) (*model.Category, error)
 	FindByName(name string) (*model.Category, error)
 	FindAll() ([]model.Category, error)
-	Update(model *model.Expense) error
-	Delete(id int) error
+	Update(model *model.Category) error
+	Delete(id string) error
 }
 
 type categoryRepository struct {
@@ -33,7 +33,7 @@ func (s *categoryRepository) NewCategory(category *model.Category) error {
 	return s.repository.Create(category)
 }
 
-func (s *categoryRepository) FindById(id int) (*model.Category, error) {
+func (s *categoryRepository) FindById(id string) (*model.Category, error) {
 	return s.repository.FindById(id)
 }
 
@@ -44,9 +44,16 @@ func (s *categoryRepository) FindByName(name string) (*model.Category, error) {
 func (s *categoryRepository) FindAll() ([]model.Category, error) {
 	return s.repository.FindAll()
 }
-func (s *categoryRepository) Update(model *model.Expense) error {
-	return s.repository.Delete(model.ID)
+func (s *categoryRepository) Update(model *model.Category) error {
+	return s.repository.Update(model)
 }
-func (s *categoryRepository) Delete(id int) error {
+func (s *categoryRepository) Delete(id string) error {
+	c, err := s.repository.FindById(id)
+	if err != nil {
+		return err
+	}
+	if c == nil {
+		return errors.New("user not found")
+	}
 	return s.repository.Delete(id)
 }
